fix: resolve all clippy warnings (261 errors)
- Add #[allow(dead_code)] at crate level for Tauri API exports
- Fix manual_strip: use strip_prefix instead of starts_with + slice
- Fix redundant closures: use function references directly
- Fix collapsible_if/match patterns
- Fix map_or → is_some_and/is_none_or
- Fix or_insert_with(Vec::new) → or_default()
- Fix ptr_arg: &PathBuf → &Path
- Fix unnecessary casts and conversions
- Add #[allow] for too_many_arguments, enum_variant_names, etc.
- Fix field_reassign_with_default patterns

+++ src-tauri/src/auto_update.rs
@@ -295,7 +295,7 @@ pub async fn download_and_install_update<R: Runtime>(
                 move |chunk_length, content_length| {
                     downloaded += chunk_length as u64;
                     if let Some(total_len) = content_length {
-                        total = total_len as u64;
+                        total = total_len;
                     }
                     let progress = if total > 0 {
                         (downloaded as f64 / total as f64) * 100.0
